// types/supplier.ts
export interface Supplier {
  id: string;
  companyName: string;
  taxCode: string;
  address: string;
  province: string;
  website: string;
  businessField: string;
  contactName: string;
  contactPosition: string;
  contactEmail: string;
  contactPhone: string;
  categories: string[];
  attachments: string[];
  status: string;
  createdAt: string;
}